Crate shams

Islamic prayer times and other solar temporal events

POSIX Time and UTC

This library uses the POSIX/UNIX time system for representing time. This was chosen to simplify working with timestamps and durations - for both myself and the user - by deferring the handling of leap seconds and synchronization with UTC to the host system.

For most use cases, the distinction between UTC and POSIX time is not significant enough to warrant consideration.

Extreme Methods

In locations towards the north and south poles, the sun may remain visible multiple days, Extreme methods are the methods that can be used to approximate (or in some cases, define) some values returned by this library. Any value that cannot be calculated in such a situation should return None.

Extreme methods need to be set to take affect.

Copyright

Copyright © 2015 Josh Morin JoshMorin@gmx.com

This program is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.

You should have received a copy of the GNU General Public License along with this program. If not, see http://www.gnu.org/licenses/.

Modules

examples
Illustrative examples of using this library.

Structs

Date
A unit structure to hold the year, month, and day expressed as 16-bit unsigned integers.

Duration
A unit i64 wrapper that represents duration expressed using Unix time.

ShamsDay
A Collection of times of interest for a specific date

ShamsSpec
Environment and Calculation Specification.

Time
A unit i64 wrapper that represents Unix time.

YearIter
Iterates over every day of a given year, yielding a ShamsDay object for each day.